changed structure of java files according devon guidelines
