--- reference-project/src/main/java/com/capgemini/academy/quarkus/domain/model/AnimalSearchCriteria.java
@@ -1,14 +1,0 @@
-package com.capgemini.academy.quarkus.domain.model;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class AnimalSearchCriteria {
-
-    private String name;
-    private Integer numberOfLegs;
-    private Integer pageNumber;
-    private Integer pageSize;
-}
