+++ reference-project/src/main/java/com/capgemini/application/quarkus/sample/animalmanagement/common/api/AnimalSearchCriteria.java
@@ -1,0 +1,17 @@
+package com.capgemini.application.quarkus.sample.animalmanagement.common.api;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class AnimalSearchCriteria {
+
+  private String name;
+
+  private Integer numberOfLegs;
+
+  private Integer pageNumber;
+
+  private Integer pageSize;
+}
