+++ reference-project/src/main/java/com/capgemini/application/quarkus/sample/animalmanagement/logic/api/to/NewAnimalDTO.java
@@ -1,0 +1,22 @@
+package com.capgemini.application.quarkus.sample.animalmanagement.logic.api.to;
+
+import org.eclipse.microprofile.openapi.annotations.enums.SchemaType;
+import org.eclipse.microprofile.openapi.annotations.media.Schema;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class NewAnimalDTO {
+
+  @Schema(nullable = false, description = "Animal name", minLength = 3, maxLength = 50)
+  private String name;
+
+  @Schema(description = "Animal tag line", minLength = 3, maxLength = 50)
+  private String basicInfo;
+
+  @Schema(type = SchemaType.INTEGER, description = "How many legs does the animal have", example = "3")
+  private int numberOfLegs;
+
+}
